apresentacao/backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
psycopg[binary]==3.2.3
python-dotenv==1.0.1
gunicorn==23.0.0
Pillow==11.0.0
whitenoise==6.8.2
